Start Rabbit Prefab
Created a new scene for the simulation and created a rabbit GameObject that uses a script that will simulates it. For now, it just defines random variables for it. I also tweaked the attributes' formats for the Animal class (for the can_eat and urge_to_run dictionaries).

Note:
For these prefabs, using parents as inputs could be useful to tweak the attributes using the parents' values.